requirements.txt
fastapi
uvicorn
pydantic
google-adk
requests
aiohttp